Update 2022_01_17_213846_create_boys_tf_tt_events.php
Add condition to BoysTimeTrialEventsTable

+++ database/migrations/2022_01_17_213846_create_boys_tf_tt_events.php
@@ -13,29 +13,31 @@ class CreateBoysTfTtEvents extends Migration
      */
     public function up()
     {
-        Schema::create('boys_tf_tt_events', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('track_time_trial_id');
-            $table->unsignedTinyInteger('track_event_id');
-            $table->unsignedTinyInteger('gender_id')->default(1)->always();
-            $table->timestamps();
+        if (!Schema::hasTable('boys_tf_tt_events')) {
+            Schema::create('boys_tf_tt_events', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('track_time_trial_id');
+                $table->unsignedTinyInteger('track_event_id');
+                $table->unsignedTinyInteger('gender_id')->default(1)->always();
+                $table->timestamps();
 
-            $table->unique(['track_time_trial_id', 'track_event_id']);
+                $table->unique(['track_time_trial_id', 'track_event_id']);
 
-            $table->foreign('track_time_trial_id')
-                ->references('id')
-                ->on('track_time_trials')
-                ->onDelete('cascade');
+                $table->foreign('track_time_trial_id')
+                    ->references('id')
+                    ->on('track_time_trials')
+                    ->onDelete('cascade');
 
-            $table->foreign('track_event_id')
-                ->references('id')
-                ->on('track_events')
-                ->onDelete('cascade');
+                $table->foreign('track_event_id')
+                    ->references('id')
+                    ->on('track_events')
+                    ->onDelete('cascade');
 
-            $table->foreign('gender_id')
-                ->references('id')
-                ->on('genders');
-        });
+                $table->foreign('gender_id')
+                    ->references('id')
+                    ->on('genders');
+            });
+        }
     }
 
     /**
